Remove verbose block and transaction constructs
With block and transaction out of BSL, we can reason about contracts
without the need to stuff inputs and outputs into transactions and
stuffing transactions into blocks. All that mundane work is handled by
the BSL exectuion engine; and the contracts now instead have the
details for the BSL engine to reason about UTXO sets.